Checks and Balances. Working on WebsocketClient.
Added some checks to make sure thrown events are handled from envManager, also moved to GlobalUsings. Musing the idea of using ANSI in the Formatter (tested on debian(putty term) and windows), changed Formatting to Environment.NewLine so on compile will use /n for *Nix and /r/n for Win. Next target for finalization is WebsocketClient.