Websocket local fault fix.
if the Websocket faulted because of the localhost losing connectivity it could get thrown into a no sleep loop of Logger  `FormatException` spamming the buffer as fast as it could. Also added a Reconnect method to slow down reconnect attempts if faulted from remote host so it doesn't try to overload the Remote, potentially causing them more issues.