Changed Logger class.
Properly renamed method to write to file as `Log`, changed the parameters to no longer require a DateTime as the first param and instead log the time at the time it is called.

Fixed unobservable error `ArgumentNullException` by converting it to a `BotException` which always reaches the buffer / logger. (Recursion protected.)